refactor(FON-468): migrate auth to feature-first architecture
