+++ apps/client/src/components/shared/ObservationDetailsContent.tsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router';
 import { getObservationDetailsPath } from '../../utils/route-path.utils';
 import { TipTapEditor } from '@/components/shared/ui/tip-tap-editor';
 import type { FilesUploader } from '@/components/shared/ui/tip-tap-editor/extensions/editor-file-uploader';
-import { useIsSg } from '@/hooks/roles.hook';
+import { useIsSg } from '@/features/auth/hooks/roles.hook';
 import { dateOnlyToDate } from '@/utils/date-only.util';
 import type { GetObservationDetailsResponseDto } from '@api/types';
