+++ apps/client/src/features/summary/components/SummaryBreadcrumb.tsx
@@ -2,9 +2,9 @@ import Button from '@codegouvfr/react-dsfr/Button';
 import { useNavigate } from 'react-router';
 
 import { Breadcrumb } from '@/components/shared/Breadcrumb';
+import { useIsSg } from '@/features/auth/hooks/roles.hook';
 import { useSummary } from '@/features/summary/context/SummaryContext';
 import { transparencyToLabel } from '@/features/transparence/labels/labels-mappers';
-import { useIsSg } from '@/hooks/roles.hook';
 import { ROUTE_PATHS } from '@/utils/route-path.utils';
 import { useDetailedNominationSessionQuery } from '@queries/nomination-sessions.queries';
