+++ apps/client/src/pages/auth/LoginPage.tsx
@@ -1,7 +1,7 @@
 import { Navigate, useLocation } from 'react-router';
 
-import Login from '@/components/login/Login';
-import { AUTHORIZED_ROLES } from '@/constants/authorized-roles.constants';
+import Login from '@/features/auth/components/Login';
+import { AUTHORIZED_ROLES } from '@/features/auth/constants/authorized-roles.constants';
 import { ROUTE_PATHS } from '@/utils/route-path.utils';
 import { useUser } from '@queries/auth.queries';
