+++ apps/client/src/pages/help/HelpPage.tsx
@@ -1,7 +1,7 @@
 import Button from '@codegouvfr/react-dsfr/Button';
 
-import { AuthGuard } from '@/components/guards/AuthGuard';
-import { AUTHORIZED_ROLES } from '@/constants/authorized-roles.constants';
+import { AuthGuard } from '@/features/auth/components/AuthGuard';
+import { AUTHORIZED_ROLES } from '@/features/auth/constants/authorized-roles.constants';
 import { ROUTE_PATHS } from '@/utils/route-path.utils';
 
 export function HelpPage() {
