+++ apps/client/src/pages/spaces/secretariat-general/SecretariatLayout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router';
 
-import { AuthGuard } from '../../../components/guards/AuthGuard';
-import { AUTHORIZED_ROLES } from '../../../constants/authorized-roles.constants';
+import { AuthGuard } from '@/features/auth/components/AuthGuard';
+import { AUTHORIZED_ROLES } from '@/features/auth/constants/authorized-roles.constants';
 
 export const SecretariatGeneralLayout = () => (
   <AuthGuard authorizedRoles={AUTHORIZED_ROLES.SG}>
